fixed pre-commit failurer

+++ inference-engine/tests/functional/inference_engine/transformations/convert_gather_to_gather_ie.cpp
@@ -25,7 +25,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEStatic1) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{15, 4, 20, 28});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{15, 4, 20, 28});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset1::Gather>(input, indices, axis_const);
 
@@ -41,7 +41,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEStatic1) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{15, 4, 20, 28});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{15, 4, 20, 28});
         auto gather = std::make_shared<op::GatherIE>(input, indices, 1);
 
         f_ref = std::make_shared<Function>(NodeVector{gather}, ParameterVector{input, indices});
@@ -55,7 +55,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEStatic2) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset1::Gather>(input, indices, axis_const);
 
@@ -71,7 +71,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEStatic2) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto unsqueeze = std::make_shared<opset1::Unsqueeze>(indices, opset1::Constant::create(element::i64, Shape{1}, {0}));
         auto gather = std::make_shared<op::GatherIE>(input, unsqueeze, 1);
         auto squeeze = std::make_shared<opset1::Squeeze>(gather, opset1::Constant::create(element::i64, Shape{1}, {1}));
@@ -87,7 +87,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEDynamic1) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN, DYN, DYN});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, PartialShape{DYN, DYN});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset1::Gather>(input, indices, axis_const);
 
@@ -102,7 +102,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEDynamic1) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN, DYN, DYN});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, PartialShape{DYN, DYN});
         auto gather = std::make_shared<op::GatherIE>(input, indices, 1);
 
         f_ref = std::make_shared<Function>(NodeVector{gather}, ParameterVector{input, indices});
@@ -116,7 +116,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEDynamic2) {
     std::shared_ptr<Function> f(nullptr), f_ref(nullptr);
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset1::Gather>(input, indices, axis_const);
 
@@ -131,7 +131,7 @@ TEST(TransformationTests, ConvertGatherToGatherIEDynamic2) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, PartialShape{DYN, DYN});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto unsqueeze = std::make_shared<opset1::Unsqueeze>(indices, opset1::Constant::create(element::i64, Shape{1}, {0}));
         auto gather = std::make_shared<op::GatherIE>(input, unsqueeze, 1);
         auto squeeze = std::make_shared<opset1::Squeeze>(gather, opset1::Constant::create(element::i64, Shape{1}, {1}));
